[bridge] Extract WorkspaceInstanceController
